rename usePlaygroundState to usePlaygroundConnectivity
