+++ src/client/src/context/playground/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from "./useCommandLine";
-export * from "./usePlaygroundState";
+export * from "./usePlaygroundConnectivity";
